+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version 3.0.20
 - Improved AOE template targeting.
+- Non PCs are marked as defeated when they drop below -10 STUN.  Once defeated they no longer get post segment 12 recoveries.
+- Improved handling of Mental attacks OMCV/DMCV, DMCV buffs and Mental Combat Skill Levels.
 
 # Version 3.0.19
 - Framework modifiers now transfer REDUCEDEND to slots [#266](https://github.com/dmdorman/hero6e-foundryvtt/issues/266)